+++ src/nested_laplace_joint_batch.cpp
@@ -443,7 +443,8 @@ Rcpp::List run_multi_block_nested_laplace_joint_batch(
                 double slope = newton_decrement(grad, st[s].delta, n_x);
                 double step = line_search_backtrack(st[s].x, st[s].delta, n_x,
                                                     obj[s], slope, eval_obj,
-                                                    obj[s], st[s].x_try);
+                                                    obj[s], st[s].x_try, nullptr,
+                                                    newton_trust_scale(conv_state[s], slope));
                 n_iter[s][kg] = iter + 1;
                 if (newton_converged(st[s].delta, grad, step, n_x, tol, conv_state[s]))
                     converged[s] = true;